[refactor] extract NotFoundError handling into helper
