+++ src/app/api/plans/[planId]/status/input/route.ts
@@ -3,7 +3,7 @@ import planInputsStatusRepo from "@/server/repos/plans/status/plan-inputs-status
 import { User } from "@supabase/supabase-js";
 import { planIdReqSchema } from "@/server/schemas/req.schema";
 import { getPlanInputsStatusResSchema } from "@/features/plan/status/server/schemas/res.schema";
-import { APIWrapper, doPlanCheck, NotFoundError } from "@/server/lib";
+import { APIWrapper, doPlanCheck, findOrThrow } from "@/server/lib/helpers";
 
 export const GET = APIWrapper(
   async (
@@ -19,10 +19,10 @@ export const GET = APIWrapper(
       data: { planId, user, permissionCheckOptions: { type: "VIEW" } },
     });
 
-    const planInputsStatus = await planInputsStatusRepo.findOne({ planId });
-    if (!planInputsStatus) {
-      throw new NotFoundError("計画入力データのステータスが見つかりません");
-    }
+    const planInputsStatus = await findOrThrow(
+      () => planInputsStatusRepo.findOne({ planId }),
+      "計画入力データのステータスが見つかりません"
+    );
 
     const res = getPlanInputsStatusResSchema.parse({
       planId,
